refactor(capabilities): rename to account/usage/get
per RFC, change name of capability
