+++ .nx/version-plans/version-plan-1758186882863.md
@@ -4,4 +4,4 @@
 '@storacha/upload-api': minor
 ---
 
-add account-usage/get capability
+add account/usage/get capability
