--- packages/upload-api/src/account-usage.js
@@ -1,6 +1,0 @@
-import { provide as provideGet } from './account-usage/get.js'
-
-/** @param {import('./types.js').AccountUsageServiceContext} context */
-export const createService = (context) => ({
-  get: provideGet(context),
-})
