--- packages/upload-api/test/handlers/account-usage.spec.js
@@ -1,3 +1,0 @@
-import * as AccountUsage from '../../src/test/handlers/account-usage.js'
-import { test } from '../../src/test/test.js'
-test({ 'account-usage/*': AccountUsage.test })
